+++ packages/editor/components/AppList/components/AppList/index.jsx
@@ -3,7 +3,15 @@ import { connect } from 'react-redux';
 import { getApps } from '../../../../actions/apps';
 import AppList from './AppList';
 
+function mapStateToProps(state) {
+  return {
+    apps: state.apps.apps,
+    error: state.apps.error,
+    user: state.user.user,
+  };
+}
+
 export default connect(
-  state => ({ apps: state.apps.apps, error: state.apps.error }),
+  mapStateToProps,
   { getApps },
 )(AppList);