+++ packages/server/api/components/responses/organization.js
@@ -1,0 +1,10 @@
+export default {
+  description: 'An organization response',
+  content: {
+    'application/json': {
+      schema: {
+        $ref: '#/components/schemas/Organization',
+      },
+    },
+  },
+};