+++ CHANGELOG.md
@@ -11,6 +11,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Block: Add support for hidden for hidden form field in form block.
 - Block: Add support for resolution limits for uploading images in form block.
+- Editor: Support basic organization management
+- Server: Support basic organization management
 
 ### Fixed